+++ packages/react/src/components/AnimatedHeader/components/AnimatedHeader/types.ts
@@ -23,4 +23,7 @@ export interface AriaLabels {
   collapseButton?: string;
   expandButton?: string;
   tilesContainer?: string;
+  carouselPrevButton?: string;
+  carouselNextButton?: string;
+  carouselDot?: string;
 }